# achieve slightly faster performance