This is refactoring to achieve slightly faster performance:

=== refactored ===
index 138398 points: 80.480ms
query 1000 closest: 4.714ms
query 50000 closest: 40.261ms
query all 138398: 84.967ms
1000 random queries of 1 closest: 52.289ms

=== original ===
index 138398 points: 80.405ms
query 1000 closest: 4.637ms
query 50000 closest: 42.415ms
query all 138398: 93.073ms
1000 random queries of 1 closest: 58.047ms