@mourner 2018-11-06
So the test fails because the distance function is no longer symmetric? (A to B isn't B to A) Or it's just that Harvesine formula is introducing some small floating point error here? Maybe we should then just compare as `Math.abs(a - b) < Number.EPSILON` or something like that instead of `a === b`.