@kkdd 2018-11-07
The following latest correction solves the symmetry problem between lat1 and lat2:
```haverSinX = Math.cos(lat1 * rad) * Math.cos(lat2 * rad) * haverSinDLng;```