@kkdd 2018-11-07
I have experimented how much it improves when avoiding sinusoidal calculations, in https://github.com/kkdd/geokdbush (i.e., ```geokdbush-braun```). I have realized, however, that it doesn't achieve significant improvement in this case.