@mourner 2018-11-07
@kkdd thanks! BTW, could you explain the math behind `vertexLat` function?