@mourner 2018-11-07
And one more quick question — in `maxHaverSinDist = haverSin(maxDistance / earthRadius)`, why do we use non-degree argument to haverSin? Or is this approximation that assumes maxDistance given is very small compared to radius? What if we supply a big one?