@kkdd 2018-11-08
Use ```\alpha_0 = \pi/2 - \phi_\text{vertex}```, ```\lambda_0 = \lambda_\text{vertex} - \pi/2``` , and ```\tan\phi = \cot\alpha_0 \sin(\lambda-\lambda_0)```, shown in https://en.wikipedia.org/wiki/Great-circle_navigation#Finding_way-points, for obtaining ```\tan\phi_\text{vertex} = \tan\phi / \cos(\lambda-\lambda_\text{vertex})```. 
> the vertex, the point on the great circle with greatest latitude