@kkdd 2018-11-08
For checking too large ```maxDistance``` assigned, the following should be required:

```js
var maxHaverSinDist = 1;
if (maxDistance !== undefined) {
    var maxCentralAngle = maxDistance / earthRadius; // in radians
    if (maxCentralAngle < Math.PI)) {
        maxHaverSinDist = haverSin(maxCentralAngle);
    }
}
```