@mourner 2018-11-08
@kkdd I was just trying to understand why the vertex point will be the closest on our longitude (producing lat extremum), but I think I figured it out — out of distances to all points on the longitude, the shortest one will arrive at a perpendicular angle to the longitude line, and by definition that point will the vertex point of a great circle line that goes from the origin point towards the longitude. Right?