@kkdd 2018-11-08
Exactly. A great circle passes perpendicularly across a longitude line, or meridian, at its vertex and it is the extremum with regard to spherical distance. It is also true in the case of geodetic lines on a spheroid.