@mourner 2018-11-08
@kkdd thanks for the explanation! Interestingly, I tried doing the same change in https://github.com/mourner/geoflatbush (which is significantly faster than geokdbush), but it doesn't seem to improve performance.