@kkdd 2018-11-09
I have also experimented and observed that the haversin version of the latest [geoflatbush](https://github.com/mourner/geoflatbush ) runs very slightly faster  (run on Macbook Pro 15 Core i7 4850HQ)

=== haversin-version ===
index 138398 points: 86.117ms
query 1000 closest: 4.490ms
query 50000 closest: 27.870ms
query all 138398: 62.811ms
2 closest for every point: 1290.604ms
100000 random queries of 1 closest: 1077.528ms

=== original ===
index 138398 points: 86.031ms
query 1000 closest: 4.646ms
query 50000 closest: 30.853ms
query all 138398: 69.355ms
2 closest for every point: 1390.765ms
100000 random queries of 1 closest: 1081.353ms