@kkdd 2018-11-09
By the way, I am now confused that the benchmark of "100000 random queries of 1 closest" of geokdbush is observed to run few-times slower than that of geoflatbush.

=== geokdbush benchmark ===
index 138398 points: 74.532ms
query 1000 closest: 4.613ms
query 50000 closest: 48.162ms
query all 138398: 81.094ms
100000 random queries of 1 closest: 3391.285ms      (*)