@mourner 2018-11-09
@kkdd the data structure of Flatbush just allows a more efficient kNN implementation that does less allocations and operates on arrays of flat numbers for the queue (as opposed to objects).